Refactor LinkModal
- Extract common helper methods into a new file (that will be imported by the UnlinkModal as well).
- Convert to a named export
- Use redux `connect` HOC to load the `actions` prop

Change-type: patch
